+++ crates/signer-ledger/src/signer.rs
@@ -2,7 +2,8 @@
 
 use crate::types::{DerivationType, LedgerError, INS, P1, P1_FIRST, P2};
 use alloy_primitives::{hex, Address};
-use alloy_signer::Signature;
+use alloy_signer::{Signature, Signer};
+use async_trait::async_trait;
 use coins_ledger::{
     common::{APDUCommand, APDUData},
     transports::{Ledger, LedgerAsync},
@@ -15,9 +16,9 @@ use futures_executor::block_on;
 #[cfg(feature = "eip712")]
 use alloy_sol_types::{Eip712Domain, SolStruct};
 
-/// A Ledger Ethereum App.
+/// A Ledger Ethereum signer.
 ///
-/// This is a simple wrapper around the [Ledger transport](Ledger)
+/// This is a simple wrapper around the [Ledger transport](Ledger).
 #[derive(Debug)]
 pub struct LedgerSigner {
     transport: Mutex<Ledger>,
@@ -33,6 +34,48 @@ impl std::fmt::Display for LedgerSigner {
             "LedgerApp. Key at index {} with address {:?} on chain_id {}",
             self.derivation, self.address, self.chain_id
         )
+    }
+}
+
+#[cfg_attr(target_arch = "wasm32", async_trait(?Send))]
+#[cfg_attr(not(target_arch = "wasm32"), async_trait)]
+impl Signer for LedgerSigner {
+    type Error = LedgerError;
+
+    async fn sign_message(&self, message: &[u8]) -> Result<Signature, Self::Error> {
+        self.sign_message(message).await
+    }
+
+    #[cfg(TODO)]
+    async fn sign_transaction(&self, message: &TypedTransaction) -> Result<Signature, Self::Error> {
+        let mut tx_with_chain = message.clone();
+        if tx_with_chain.chain_id().is_none() {
+            // in the case we don't have a chain_id, let's use the signer chain id instead
+            tx_with_chain.set_chain_id(self.chain_id);
+        }
+        self.sign_tx(&tx_with_chain).await
+    }
+
+    #[cfg(feature = "eip712")]
+    async fn sign_typed_data<T: SolStruct + Send + Sync>(
+        &self,
+        payload: &T,
+        domain: &Eip712Domain,
+    ) -> Result<Signature, Self::Error> {
+        self.sign_typed_struct(payload, domain).await
+    }
+
+    fn address(&self) -> Address {
+        self.address
+    }
+
+    fn with_chain_id<T: Into<u64>>(mut self, chain_id: T) -> Self {
+        self.chain_id = chain_id.into();
+        self
+    }
+
+    fn chain_id(&self) -> u64 {
+        self.chain_id
     }
 }
